feat: use and refactor DisclaimerMessage for the DonationComponent
diff --git a/packages/lib/src/components/Card/components/CardInput/components/CardFieldsWrapper.tsx b/packages/lib/src/components/Card/components/CardInput/components/CardFieldsWrapper.tsx
index e9195976..a20037c7 100644
--- a/packages/lib/src/components/Card/components/CardInput/components/CardFieldsWrapper.tsx
+++ b/packages/lib/src/components/Card/components/CardInput/components/CardFieldsWrapper.tsx
@@ -7,7 +7,7 @@ import StoreDetails from '../../../../internal/StoreDetails';
 import Address from '../../../../internal/Address';
 import CardHolderName from './CardHolderName';
 import Installments from './Installments';
-import DisclaimerMessage from './DisclaimerMessage';
+import DisclaimerMessage from '../../../../internal/DisclaimerMessage';

 export const CardFieldsWrapper = ({
     // vars created in CardInput:
diff --git a/packages/lib/src/components/Card/components/CardInput/components/StoredCardFieldsWrapper.tsx b/packages/lib/src/components/Card/components/CardInput/components/StoredCardFieldsWrapper.tsx
index de2fbbb6..be2b5521 100644
--- a/packages/lib/src/components/Card/components/CardInput/components/StoredCardFieldsWrapper.tsx
+++ b/packages/lib/src/components/Card/components/CardInput/components/StoredCardFieldsWrapper.tsx
@@ -2,7 +2,7 @@ import { h } from 'preact';
 import LoadingWrapper from '../../../../internal/LoadingWrapper';
 import StoredCardFields from './StoredCardFields';
 import Installments from './Installments';
-import DisclaimerMessage from './DisclaimerMessage';
+import DisclaimerMessage from '../../../../internal/DisclaimerMessage';

 export const StoredCardFieldsWrapper = ({
     // base (shared)
diff --git a/packages/lib/src/components/Card/components/CardInput/types.ts b/packages/lib/src/components/Card/components/CardInput/types.ts
index fece0aae..5227ea9a 100644
--- a/packages/lib/src/components/Card/components/CardInput/types.ts
+++ b/packages/lib/src/components/Card/components/CardInput/types.ts
@@ -11,6 +11,7 @@ import { SRPanel } from '../../../../core/Errors/SRPanel';
 import Analytics from '../../../../core/Analytics';
 import RiskElement from '../../../../core/RiskModule';
 import { ComponentMethodsRef } from '../../../types';
+import { DisclaimerMsgObject } from '../../../internal/DisclaimerMessage/DisclaimerMessage';

 export interface CardInputValidState {
     holderName?: boolean;
@@ -46,12 +47,6 @@ type Placeholders = {
     holderName?: string;
 };

-export interface DisclaimerMsgObject {
-    message: string;
-    linkText: string;
-    link: string;
-}
-
 /**
  * Should be the subset of the props sent to CardInput that are *actually* used by CardInput
  * - either in the comp itself or are passed on to its children
diff --git a/packages/lib/src/components/Donation/Donation.scss b/packages/lib/src/components/Donation/Donation.scss
index 0897819e..8a84d9dc 100644
--- a/packages/lib/src/components/Donation/Donation.scss
+++ b/packages/lib/src/components/Donation/Donation.scss
@@ -1,4 +1,4 @@
-@import "../../style/index";
+@import '../../style/index';

 .adyen-checkout__adyen-giving {
     .adyen-checkout__status__icon {
@@ -92,9 +92,16 @@
 }

 .adyen-checkout__adyen-giving-actions {
+    display: flex;
+    flex-direction: column;
     margin-top: $spacing-medium;
+
+    .adyen-checkout-disclaimer__label {
+        align-self: center;
+    }
 }

+
 .adyen-checkout__button {
     &.adyen-checkout__button--donate {
         margin: $spacing-medium auto $spacing-small;
diff --git a/packages/lib/src/components/Donation/components/DonationComponent.tsx b/packages/lib/src/components/Donation/components/DonationComponent.tsx
index 15a23c56..b2146ff4 100644
--- a/packages/lib/src/components/Donation/components/DonationComponent.tsx
+++ b/packages/lib/src/components/Donation/components/DonationComponent.tsx
@@ -7,9 +7,9 @@ import Img from '../../internal/Img';
 import { getImageUrl } from '../../../utils/get-image';
 import useCoreContext from '../../../core/Context/useCoreContext';
 import '../Donation.scss';
-
+import DisclaimerMessage from '../../internal/DisclaimerMessage';
 export default function DonationComponent(props) {
-    const { amounts, onCancel, onDonate, showCancelButton = true } = props;
+    const { amounts, onCancel, onDonate, showCancelButton = true, disclaimerMessage } = props;
     const { i18n, loadingContext } = useCoreContext();
     const { currency } = amounts;
     const [status, setStatus] = useState('ready');
@@ -89,7 +89,7 @@ export default function DonationComponent(props) {
                         onChange={handleAmountSelected}
                     />
                 </div>
-
+                {disclaimerMessage && <DisclaimerMessage disclaimer={disclaimerMessage} />}
                 <Button
                     classNameModifiers={['donate']}
                     onClick={handleDonate}
diff --git a/packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/DisclaimerMessage.scss b/packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.scss
similarity index 78%
rename from packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/DisclaimerMessage.scss
rename to packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.scss
index 95484072..6dac2fbc 100644
--- a/packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/DisclaimerMessage.scss
+++ b/packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.scss
@@ -1,16 +1,16 @@
-@import "../../../../../../style/index";
+@import "../../../style/index";

 .adyen-checkout-disclaimer__label {
     position: relative;
     margin-top: $spacing-medium;
     padding-left: 0;
-    cursor: pointer;
+    cursor: auto;
     display: inline-block;
     line-height: 19px;
     color: $color-black;
     font-size: $font-size-small;
     font-weight: normal;
-    user-select: none;
+

     [dir="rtl"] & {
         padding-right: 0;
diff --git a/packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/DisclaimerMessage.test.tsx b/packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.test.tsx
similarity index 100%
rename from packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/DisclaimerMessage.test.tsx
rename to packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.test.tsx
diff --git a/packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/DisclaimerMessage.tsx b/packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.tsx
similarity index 88%
rename from packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/DisclaimerMessage.tsx
rename to packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.tsx
index 2a42491c..54ade3a1 100644
--- a/packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/DisclaimerMessage.tsx
+++ b/packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.tsx
@@ -1,8 +1,13 @@
 import { h } from 'preact';
-import { DisclaimerMsgObject } from '../../types';
-import { isValidHttpUrl } from '../../../../../../utils/isValidURL';
+import { isValidHttpUrl } from '../../../utils/isValidURL';
 import './DisclaimerMessage.scss';

+export interface DisclaimerMsgObject {
+    message: string;
+    linkText: string;
+    link: string;
+}
+
 interface DisclaimerMessageProps {
     disclaimer: DisclaimerMsgObject;
 }
diff --git a/packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/index.ts b/packages/lib/src/components/internal/DisclaimerMessage/index.ts
similarity index 100%
rename from packages/lib/src/components/Card/components/CardInput/components/DisclaimerMessage/index.ts
rename to packages/lib/src/components/internal/DisclaimerMessage/index.ts
