+++ packages/lib/src/components/internal/DisclaimerMessage/DisclaimerMessage.scss
@@ -1,16 +1,16 @@
-@import "../../../../../../style/index";
+@import "../../../style/index";
 
 .adyen-checkout-disclaimer__label {
     position: relative;
     margin-top: $spacing-medium;
     padding-left: 0;
-    cursor: pointer;
+    cursor: auto;
     display: inline-block;
     line-height: 19px;
     color: $color-black;
     font-size: $font-size-small;
     font-weight: normal;
-    user-select: none;
+
 
     [dir="rtl"] & {
         padding-right: 0;
